// administradores-microservice/src/shared/enums/error-codes.enum.ts
export enum ErrorCodesEnum {
  INTERNAL_ERROR = 'internal_error',
  UNKNOW_ERROR = 'unknow_error',
  UNAUTHORIZED = 'unauthorized',
  INVALID_ID = 'invalid_id',
  ADMIN_NOT_FOUND = 'admin_not_found',
  INVALID_BIRTH_DATE = 'invalid_birth_date',
  INVALID_CPF = 'invalid_cpf',
  INVALID_EMAIL = 'invalid_email',
  REQUIRED_NAME = 'required_name',
  REQUIRED_BIRTH_DATE = 'required_birth_date',
  REQUIRED_CRM = 'required_crm',
  PATIENT_NOT_FOUND = 'patient_not_found',
  PATIENT_ALREADY_EXIST = 'patient_already_exist',
  INVALID_DOCUMENT = 'invalid_document',
  INVALID_STATE = 'invalid_state',
  INVALID_FORMAT_PHONE_NUMBER = 'invalid_format_phone_number',
}