// administradores-microservice/src/shared/enums/error-messages.enum.ts
export enum ErrorMessagesEnum {
  INTERNAL_ERROR = 'Erro interno no servidor.',
  UNKNOW_ERROR = 'Erro desconhecido.',
  UNAUTHORIZED = 'Unauthorized.',
  INVALID_ID = 'O id informado é inválido.',
  ADMIN_NOT_FOUND = 'Administrador inválido.',
  INVALID_BIRTH_DATE = 'invalid_birth_date',
  INVALID_CPF = 'invalid_cpf',
  INVALID_EMAIL = 'invalid_email',
  REQUIRED_NAME = 'required_name',
  REQUIRED_BIRTH_DATE = 'required_birth_date',
  REQUIRED_CRM = 'required_crm',
  PATIENT_NOT_FOUND = 'patient_not_found',
  PATIENT_ALREADY_EXIST = 'patient_already_exist',
  INVALID_DOCUMENT = 'invalid_document',
  INVALID_STATE = 'invalid_state',
  INVALID_FORMAT_PHONE_NUMBER = 'invalid_format_phone_number',
}